pyproject.toml
[project]
name = "excel-operation"
version = "0.1.0"
description = ""
authors = [
    {name = "akitorahayashi",email = "atora.has75@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "pandas (>=2.0.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
